backend/requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.9
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==4.0
django-allauth==0.47.0
django-cors-headers==3.10.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
oauthlib==3.1.1
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
zope.interface==5.4.0
